+++ examples/ring-malli-swagger/src/example/server.clj
@@ -61,7 +61,7 @@
          {:get {:summary "downloads a file"
                 :swagger {:produces ["image/png"]}
                 :responses {200 {:description "an image"
-                                 :content {"image/png" {:schema any?}}}}
+                                 :content {"image/png" {:schema string?}}}}
                 :handler (fn [_]
                            {:status 200
                             :headers {"Content-Type" "image/png"}